Convert with parallel preprocessing of files
